debug itoa

+++ includes/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: mcherel- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/07 23:22:48 by mcherel-          #+#    #+#             */
-/*   Updated: 2022/01/16 18:54:31 by mcherel-         ###   ########.fr       */
+/*   Updated: 2022/01/17 13:56:02 by mcherel-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ char	*ft_itoa(int n);
 //void    *ft_converts(const char *str);//en cours
 size_t	ft_strlen( const char *str );
 void    ft_putchar(char c);
-int     ft_getnblen(int nb, int baselen);
+int     ft_getnblen(long int nb, int baselen);
 char	*ft_nbtostr(char *str, unsigned int nb, int len, char *base);
 char	*ft_utoa(unsigned int n);
 char    *ft_xtoa(int n);
