+++ srcs/ft_getnblen.c
@@ -6,13 +6,13 @@
 /*   By: mcherel- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/12 08:09:55 by mcherel-          #+#    #+#             */
-/*   Updated: 2022/01/14 14:47:03 by mcherel-         ###   ########.fr       */
+/*   Updated: 2022/01/17 14:43:41 by mcherel-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-int     ft_getnblen(size_t nb, int baselen)
+int     ft_getnblen(long int nb, int baselen)
 {
 	int	i;
